ensure pid exists

+++ lib/Thruk.pm
@@ -109,7 +109,7 @@ Thruk::Utils::IO::mkdir(__PACKAGE__->config->{'tmp_path'});
 my $pidfile  = __PACKAGE__->config->{'tmp_path'}.'/thruk.pid';
 sub _remove_pid {
     if(defined $ENV{'THRUK_SRC'} and $ENV{'THRUK_SRC'} eq 'FastCGI') {
-        if(-f $pidfile) {
+        if($pidfile && -f $pidfile) {
             my $pid = read_file($pidfile);
             chomp($pid);
             unlink($pidfile) if $pid == $$;
